feat: get statictis for guide and tour

+++ src/main/java/com/example/localguidebe/controller/GuideController.java
@@ -44,7 +44,7 @@ public class GuideController {
                 new SearchGuideDTO(
                     guides.stream().map(userToGuideDtoConverter::convert).toList(),
                     guides.getTotalPages(),
-                    guides.getTotalElements())));
+                    (int) guides.getTotalElements())));
   }
 
   @GetMapping("/{id}")
